+++ types/downtimes/database.go
@@ -3,7 +3,6 @@ package downtimes
 import (
 	"fmt"
 	"github.com/statping/statping/database"
-	"github.com/statping/statping/utils"
 	"strconv"
 	"time"
 )
@@ -60,11 +59,14 @@ func FindByService(service int64, start time.Time, end time.Time) (*[]Downtime, 
 	return &downtime, q.Error()
 }
 
-func FindDowntime(timeVar time.Time) []Downtime {
-	var downtime []Downtime
-	q := db.Where("start <= ? and \"end\" >= ?", timeVar, timeVar)
-	q = q.Order("id ASC").Find(&downtime)
-	return downtime
+func ConvertToUnixTime(str string) (time.Time, error) {
+	i, err := strconv.ParseInt(str, 10, 64)
+	var t time.Time
+	if err != nil {
+		return t, err
+	}
+	tm := time.Unix(i, 0)
+	return tm, nil
 }
 
 func FindAll(vars map[string]string) (*[]Downtime, error) {
@@ -76,11 +78,11 @@ func FindAll(vars map[string]string) (*[]Downtime, error) {
 	startInt, err := strconv.ParseInt(st, 10, 64)
 	endInt, err := strconv.ParseInt(en, 10, 64)
 	if err1 && err2 && (endInt > startInt) {
-		start, err = utils.ConvertToUnixTime(vars["start"])
+		start, err = ConvertToUnixTime(vars["start"])
 		if err != nil {
 			return &downtime, err
 		}
-		end, err = utils.ConvertToUnixTime(vars["end"])
+		end, err = ConvertToUnixTime(vars["end"])
 		if err != nil {
 			return &downtime, err
 		}
@@ -118,7 +120,6 @@ func FindAll(vars map[string]string) (*[]Downtime, error) {
 	q = q.Limit((int)(count)).Offset((int)(skip)).Find(&downtime)
 	return &downtime, q.Error()
 }
-
 func (c *Downtime) Create() error {
 	q := db.Create(c)
 	return q.Error()
